+++ cluster_tools/cluster_tools/util.py
@@ -162,7 +162,7 @@ class FileWaitThread(threading.Thread):
                         # Check for output file as a fast indicator for job completion
                         handle_completed_job(job_id, filename, False)
                     elif self.executor is not None:
-                        status = self.executor.check_for_crashed_job(job_id)
+                        status = self.executor.check_job_state(job_id)
 
                         # We have to re-check for the output file since this could be created in the mean time
                         if os.path.exists(filename):