+++ webknossos/examples/learned_segmenter.py
@@ -16,14 +16,13 @@ def main() -> None:
     # We are going to use a public demo annotation for this example
 
     annotation = wk.Annotation.download(
-        "https://webknossos.org/annotations/Explorational/616457c2010000870032ced4"
+        "https://webknossos.org/annotations/616457c2010000870032ced4"
     )
 
-    # Step 1: Download the dataset and our training data annotation from webKnossos to our local computer
+    # Step 1: Download the training data annotation from webKnossos to our local computer
     training_data_bbox = annotation.user_bounding_boxes[0]  # type: ignore[index]
     time_str = strftime("%Y-%m-%d_%H-%M-%S", gmtime())
     new_dataset_name = annotation.dataset_name + f"_segmented_{time_str}"
-    mag = wk.Mag(1)
     dataset = wk.Dataset.open_remote(
         annotation.dataset_name,
         organization_id="scalable_minds",
@@ -31,11 +30,12 @@ def main() -> None:
     )
 
     with annotation.temporary_volume_layer_copy() as volume_annotation_layer:
+        mag = volume_annotation_layer.get_finest_mag().mag
         volume_annotation_data = volume_annotation_layer.mags[mag].read(
             absolute_bounding_box=training_data_bbox
         )
 
-    mag_view = dataset.layers["color"].mags[mag]
+    color_mag_view = dataset.layers["color"].mags[mag]
 
     # Step 2: Initialize a machine learning model to segment our dataset
     features_func = partial(
@@ -46,7 +46,7 @@ def main() -> None:
     # Step 3: Manipulate our data to fit the ML model and start training on
     # data from our annotated training data bounding box
     print("Starting training…")
-    img_data_train = mag_view.read(
+    img_data_train = color_mag_view.read(
         absolute_bounding_box=training_data_bbox
     )  # wk data has dimensions (Channels, X, Y, Z)
     # move channels to last dimension, remove z dimension to match skimage's shape
@@ -58,7 +58,7 @@ def main() -> None:
     # Step 4: Use our trained model and predict a class for each pixel in the dataset
     # to get a full segmentation of the data
     print("Starting prediction…")
-    X_predict = np.moveaxis(np.squeeze(mag_view.read()), 0, -1)
+    X_predict = np.moveaxis(np.squeeze(color_mag_view.read()), 0, -1)
     Y_predicted = segmenter.predict(X_predict)
     segmentation = Y_predicted[:, :, None]  # adds z dimension
     assert segmentation.max() < 256
@@ -79,19 +79,13 @@ def main() -> None:
         segmentation_layer.bounding_box = dataset.layers["color"].bounding_box
         segmentation_layer.add_mag(mag, compress=True).write(segmentation)
 
-        uploaded_dataset = new_dataset.upload(
-            layers_to_link=[
-                wk.LayerToLink(
-                    organization_id="scalable_minds",
-                    dataset_name=annotation.dataset_name,
-                    layer_name="color",
-                )
-            ]
+        remote_ds = dataset.upload(
+            layers_to_link=[annotation.get_remote_base_dataset().get_layer("color")]
             if "PYTEST_CURRENT_TEST" not in os.environ
             else None
         )
 
-    url = uploaded_dataset.url
+    url = remote_ds.url
     print(f"Successfully uploaded {url}")