+++ webknossos/webknossos/annotation/annotation.py
@@ -49,6 +49,7 @@ from zipp import Path as ZipPath
 import webknossos._nml as wknml
 from webknossos.annotation._nml_conversion import annotation_to_nml, nml_to_skeleton
 from webknossos.dataset import SEGMENTATION_CATEGORY, Dataset, Layer, SegmentationLayer
+from webknossos.dataset.dataset import RemoteDataset
 from webknossos.geometry import BoundingBox
 from webknossos.skeleton import Skeleton
 from webknossos.utils import time_since_epoch_in_ms, warn_deprecated
@@ -58,8 +59,8 @@ Vector3 = Tuple[float, float, float]
 
 MAG_RE = r"((\d+-\d+-)?\d+)"
 SEP_RE = r"(\/|\\)"
-CUBE_RE = fr"z\d+{SEP_RE}y\d+{SEP_RE}x\d+\.wkw"
-ANNOTATION_WKW_PATH_RE = re.compile(fr"{MAG_RE}{SEP_RE}(header\.wkw|{CUBE_RE})")
+CUBE_RE = rf"z\d+{SEP_RE}y\d+{SEP_RE}x\d+\.wkw"
+ANNOTATION_WKW_PATH_RE = re.compile(rf"{MAG_RE}{SEP_RE}(header\.wkw|{CUBE_RE})")
 
 
 @attr.define
@@ -83,7 +84,7 @@ class Annotation:
     _voxel_size: Optional[Vector3] = None
     _organization_id: Optional[str] = None
     _description: Optional[str] = None
-    username: Optional[str] = None
+    owner_name: Optional[str] = None
     annotation_id: Optional[str] = None
     time: Optional[int] = attr.ib(factory=time_since_epoch_in_ms)
     edit_position: Optional[Vector3] = None
@@ -134,6 +135,17 @@ class Annotation:
                 "When supplying a skeleton for Annotation(), passing dataset_name, voxel_size, organization_id or description is not allowed. "
                 + "The attributes of the skeleton are used in this case."
             )
+
+    @property
+    def username(self) -> Optional[str]:
+        """Deprecated, use `owner_name` instead."""
+        warn_deprecated("username", "owner_name")
+        return self.owner_name
+
+    @username.setter
+    def username(self, username: str) -> None:
+        warn_deprecated("username", "owner_name")
+        self.owner_name = username
 
     @property
     def dataset_name(self) -> str:
@@ -217,8 +229,8 @@ class Annotation:
     ) -> "Annotation":
         """
         * `annotation_id_or_url` may be an annotation id or a full URL to an annotation, e.g.
-          `https://webknossos.org/annotations/Explorational/6114d9410100009f0096c640`
-        * `annotation_type` must be supplied if and only if an annotation id was used in the previous argument
+          `https://webknossos.org/annotations/6114d9410100009f0096c640`
+        * `annotation_type` is no longer required and therefore deprecated and ignored
         * `webknossos_url` may be supplied if an annotation id was used
           and allows to specifiy in which webknossos instance to search for the annotation.
           It defaults to the url from your current `webknossos_context`, using https://webknossos.org as a fallback.
@@ -234,23 +246,19 @@ class Annotation:
         if match is not None:
             assert webknossos_url is None and annotation_type is None, (
                 "When Annotation.download() is be called with an annotation url, "
-                + "e.g. Annotation.download('https://webknossos.org/annotations/Explorational/6114d9410100009f0096c640'), "
+                + "e.g. Annotation.download('https://webknossos.org/annotations/6114d9410100009f0096c640'), "
                 + "annotation_type and webknossos_url must not be set."
             )
             annotation_id = match.group("annotation_id")
-            annotation_type = match.group("annotation_type")
             webknossos_url = match.group("webknossos_url")
         else:
-            assert annotation_type is not None, (
-                "When calling Annotation.download() with an id you must supply the argument annotation_type, "
-                + "e.g. 'Task' or 'Explorational'. Alternatively, you can use the full annotation url, "
-                + "e.g. Annotation.download('https://webknossos.org/annotations/Explorational/6114d9410100009f0096c640')."
-            )
             annotation_id = annotation_id_or_url
-        annotation_type = AnnotationType(annotation_type)
-        assert (
-            annotation_type not in _COMPOUND_ANNOTATION_TYPES
-        ), f"Currently compound annotation types are not supported, got {annotation_type}"
+
+        if annotation_type is not None:
+            warnings.warn(
+                "[DEPRECATION] `annotation_type` is deprecated for Annotation.download(), it should be omitted.",
+                DeprecationWarning,
+            )
 
         if webknossos_url is not None and webknossos_url != _get_context().url:
             warnings.warn(
@@ -267,7 +275,7 @@ class Annotation:
         with context:
             client = _get_generated_client()
             response = annotation_download.sync_detailed(
-                typ=annotation_type.value, id=annotation_id, client=client
+                id=annotation_id, client=client
             )
         assert response.status_code == 200, response
         content_disposition_header = response.headers.get("content-disposition", "")
@@ -297,7 +305,7 @@ class Annotation:
             cls(
                 name=name,
                 skeleton=nml_to_skeleton(nml),
-                username=nml.get_meta("username"),
+                owner_name=nml.get_meta("username"),
                 annotation_id=nml.get_meta("annotationId"),
                 time=nml.parameters.time,
                 edit_position=nml.parameters.editPosition,
@@ -425,6 +433,9 @@ class Annotation:
                 volume_layer._default_zip_name(),
                 layer_content,
             )
+
+    def get_remote_base_dataset(self) -> RemoteDataset:
+        return Dataset.open_remote(self.dataset_name, self.organization_id)
 
     def get_volume_layer_names(self) -> Iterable[str]:
         return (i.name for i in self._volume_layers)
@@ -626,6 +637,8 @@ class Annotation:
                 volume_layer_id=volume_layer_id,
             )
 
+            input_annotation_dataset._read_only = True
+
             yield input_annotation_layer
 
 
@@ -674,7 +687,7 @@ _COMPOUND_ANNOTATION_TYPES = [
 
 _ANNOTATION_URL_REGEX = re.compile(
     r"^(?P<webknossos_url>https?://.*)/annotations/"
-    + fr"(?P<annotation_type>{'|'.join(i.value for i in AnnotationType.__members__.values())})/"
+    + rf"((?P<annotation_type>{'|'.join(i.value for i in AnnotationType.__members__.values())})/)?"
     + r"(?P<annotation_id>[0-9A-Fa-f]*)"
 )