+++ webknossos/webknossos/client/_generated/models/task_infos_by_project_id_response_200_item.py
@@ -1,4 +1,4 @@
-from typing import Any, Dict, List, Optional, Type, TypeVar, cast
+from typing import Any, Dict, List, Optional, Type, TypeVar, Union, cast
 
 import attr
 
@@ -11,6 +11,7 @@ from ..models.task_infos_by_project_id_response_200_item_status import (
 from ..models.task_infos_by_project_id_response_200_item_type import (
     TaskInfosByProjectIdResponse200ItemType,
 )
+from ..types import UNSET, Unset
 
 T = TypeVar("T", bound="TaskInfosByProjectIdResponse200Item")
 
@@ -28,13 +29,13 @@ class TaskInfosByProjectIdResponse200Item:
     data_set: str
     needed_experience: TaskInfosByProjectIdResponse200ItemNeededExperience
     created: int
-    status: TaskInfosByProjectIdResponse200ItemStatus
     script: str
     creation_info: str
     bounding_box: str
     edit_position: List[int]
     edit_rotation: List[int]
     tracing_time: Optional[int]
+    status: Union[Unset, TaskInfosByProjectIdResponse200ItemStatus] = UNSET
     additional_properties: Dict[str, Any] = attr.ib(init=False, factory=dict)
 
     def to_dict(self) -> Dict[str, Any]:
@@ -49,14 +50,16 @@ class TaskInfosByProjectIdResponse200Item:
         needed_experience = self.needed_experience.to_dict()
 
         created = self.created
-        status = self.status.to_dict()
-
         script = self.script
         creation_info = self.creation_info
         bounding_box = self.bounding_box
         edit_position = self.edit_position
 
         edit_rotation = self.edit_rotation
+
+        status: Union[Unset, Dict[str, Any]] = UNSET
+        if not isinstance(self.status, Unset):
+            status = self.status.to_dict()
 
         tracing_time = self.tracing_time
 
@@ -73,7 +76,6 @@ class TaskInfosByProjectIdResponse200Item:
                 "dataSet": data_set,
                 "neededExperience": needed_experience,
                 "created": created,
-                "status": status,
                 "script": script,
                 "creationInfo": creation_info,
                 "boundingBox": bounding_box,
@@ -82,6 +84,8 @@ class TaskInfosByProjectIdResponse200Item:
                 "tracingTime": tracing_time,
             }
         )
+        if status is not UNSET:
+            field_dict["status"] = status
 
         return field_dict
 
@@ -110,8 +114,6 @@ class TaskInfosByProjectIdResponse200Item:
 
         created = d.pop("created")
 
-        status = TaskInfosByProjectIdResponse200ItemStatus.from_dict(d.pop("status"))
-
         script = d.pop("script")
 
         creation_info = d.pop("creationInfo")
@@ -121,6 +123,13 @@ class TaskInfosByProjectIdResponse200Item:
         edit_position = cast(List[int], d.pop("editPosition"))
 
         edit_rotation = cast(List[int], d.pop("editRotation"))
+
+        _status = d.pop("status", UNSET)
+        status: Union[Unset, TaskInfosByProjectIdResponse200ItemStatus]
+        if isinstance(_status, Unset):
+            status = UNSET
+        else:
+            status = TaskInfosByProjectIdResponse200ItemStatus.from_dict(_status)
 
         tracing_time = d.pop("tracingTime")
 
@@ -134,12 +143,12 @@ class TaskInfosByProjectIdResponse200Item:
             data_set=data_set,
             needed_experience=needed_experience,
             created=created,
-            status=status,
             script=script,
             creation_info=creation_info,
             bounding_box=bounding_box,
             edit_position=edit_position,
             edit_rotation=edit_rotation,
+            status=status,
             tracing_time=tracing_time,
         )