+++ webknossos/webknossos/client/_resumable/chunk.py
@@ -1,4 +1,5 @@
 import mimetypes
+import time
 from pathlib import Path
 from typing import Any, Dict
 
@@ -6,6 +7,8 @@ import httpx
 
 from .file import FileChunk, ResumableFile
 from .util import Config
+
+RETRY_DELAY = 5  # seconds
 
 
 class ResumableError(Exception):
@@ -36,6 +39,7 @@ def resolve_chunk(
     if not exists_on_server:
         tries = 0
         while not _send_chunk(client, config, file, chunk):
+            time.sleep(RETRY_DELAY)
             tries += 1
             if tries >= config.max_chunk_retries:
                 raise ResumableError("max retries exceeded")
@@ -74,11 +78,15 @@ def _send_chunk(
     ResumableError
         If the server responded with an error code indicating permanent failure
     """
-    response = client.post(
-        config.target,
-        data=_build_query(file, chunk, config.additional_query_params),
-        files={"file": chunk.read()},
-    )
+    try:
+        response = client.post(
+            config.target,
+            data=_build_query(file, chunk, config.additional_query_params),
+            files={"file": chunk.read()},
+        )
+    except httpx.TransportError:
+        # The request itself failed. The calling function will retry.
+        return False
     if response.status_code in config.permanent_errors:
         raise ResumableError(f"{response.status_code} Error: {response.text}")
     return response.status_code in [200, 201]