+++ webknossos/webknossos/client/context.py
@@ -123,7 +123,7 @@ def _clear_all_context_caches() -> None:
 
 @attr.frozen
 class _WebknossosContext:
-    url: str = os.environ.get("WK_URL", default=DEFAULT_WEBKNOSSOS_URL)
+    url: str = os.environ.get("WK_URL", default=DEFAULT_WEBKNOSSOS_URL).rstrip("/")
     token: Optional[str] = os.environ.get("WK_TOKEN", default=None)
     timeout: int = int(os.environ.get("WK_TIMEOUT", default=DEFAULT_HTTP_TIMEOUT))
 
@@ -132,7 +132,14 @@ class _WebknossosContext:
     @property
     def required_token(self) -> str:
         if self.token is None:
-            return _cached_ask_for_token(self.url)
+            token = _cached_ask_for_token(self.url)
+            # We replace the current context, but leave all previous ones as-is.
+            # Any opened contextmanagers will still close correctly, as the stored
+            # tokens still point to the correct predecessors.
+            _webknossos_context_var.set(
+                _WebknossosContext(self.url, token, self.timeout)
+            )
+            return token
         else:
             return self.token
 
@@ -204,7 +211,7 @@ class webknossos_context(ContextDecorator):
         `url` and `timeout` are taken from the previous context (e.g. environment variables) if not specified.
         `token` must be set explicitly, it is not available when not specified.
         """
-        self._url = _get_context().url if url is None else url
+        self._url = _get_context().url if url is None else url.rstrip("/")
         self._token = token
         self._timeout = _get_context().timeout if timeout is None else timeout
         self._context_var_token_stack: List[Token[_WebknossosContext]] = []